requirements.txt
python-telegram-bot[job-queue]==21.3
psycopg2-binary
requests~=2.32.3
beautifulsoup4~=4.12.3
telegram